Move FailableStream and its tests to stream package.
